+++ sazanov.maxim/T2/DataStruct.cpp
@@ -3,7 +3,7 @@
 #include <bitset>
 #include <DelimiterI.hpp>
 #include <StreamGuard.hpp>
-#include <ValueIO.hpp>
+#include "ValueIO.hpp"
 
 std::istream& sazanov::operator>>(std::istream& in, DataStruct& value)
 {
