+++ sazanov.maxim/T2/ValueIO.hpp
@@ -17,12 +17,6 @@ namespace sazanov
   };
   std::ostream& operator<<(std::ostream& out, const BinUllO&& value);
 
-  struct DoubleO
-  {
-    const double& number;
-  };
-  std::ostream& operator<<(std::ostream& out, const DoubleO&& value);
-
   struct CharKeyI
   {
     char& ch;
